fix: move relocation procecure to BookKeeperAdmin
