+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/tools/cli/commands/bookies/CorrectEnsemblePlacementCommand.java
@@ -20,20 +20,15 @@ package org.apache.bookkeeper.tools.cli.commands.bookies;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.converters.CommaParameterSplitter;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Functions;
 import com.google.common.util.concurrent.UncheckedExecutionException;
 import java.io.IOException;
 import java.net.URI;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.NavigableSet;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.CountDownLatch;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import lombok.Cleanup;
 import lombok.Setter;
 import lombok.experimental.Accessors;
@@ -41,9 +36,6 @@ import org.apache.bookkeeper.bookie.BookieException;
 import org.apache.bookkeeper.client.BKException;
 import org.apache.bookkeeper.client.BookKeeper;
 import org.apache.bookkeeper.client.BookKeeperAdmin;
-import org.apache.bookkeeper.client.EnsemblePlacementPolicy;
-import org.apache.bookkeeper.client.LedgerFragment;
-import org.apache.bookkeeper.client.api.LedgerMetadata;
 import org.apache.bookkeeper.conf.ClientConfiguration;
 import org.apache.bookkeeper.conf.ServerConfiguration;
 import org.apache.bookkeeper.meta.LedgerManagerFactory;
@@ -51,7 +43,6 @@ import org.apache.bookkeeper.meta.LedgerUnderreplicationManager;
 import org.apache.bookkeeper.meta.MetadataBookieDriver;
 import org.apache.bookkeeper.meta.MetadataDrivers;
 import org.apache.bookkeeper.meta.exceptions.MetadataException;
-import org.apache.bookkeeper.net.BookieId;
 import org.apache.bookkeeper.replication.ReplicationException;
 import org.apache.bookkeeper.stats.NullStatsLogger;
 import org.apache.bookkeeper.tools.cli.helpers.BookieCommand;
@@ -133,7 +124,7 @@ public class CorrectEnsemblePlacementCommand extends
 
             final ClientConfiguration clientConf = new ClientConfiguration(conf);
             final BookKeeper bookKeeper = new BookKeeper(clientConf);
-            final BookKeeperAdmin admin = new BookKeeperAdmin(clientConf);
+            final BookKeeperAdmin admin = new BookKeeperAdmin(bookKeeper);
             return relocate(conf, flags, bookKeeper, admin);
         } catch (Exception e) {
             throw new UncheckedExecutionException(e.getMessage(), e);
@@ -143,8 +134,6 @@ public class CorrectEnsemblePlacementCommand extends
     @VisibleForTesting
     public boolean relocate(ServerConfiguration conf, CorrectEnsemblePlacementFlags flags,
                             BookKeeper bookKeeper, BookKeeperAdmin admin) throws Exception {
-        final EnsemblePlacementPolicy placementPolicy = bookKeeper.getPlacementPolicy();
-
         @Cleanup
         final MetadataBookieDriver metadataDriver = instantiateMetadataDriver(conf);
         @Cleanup
@@ -153,7 +142,7 @@ public class CorrectEnsemblePlacementCommand extends
         final LedgerUnderreplicationManager lum = lmf.newLedgerUnderreplicationManager();
 
         final List<Long> targetLedgers =
-                flags.ledgerIds.stream().parallel().distinct().filter(ledgerId -> {
+                flags.ledgerIds.stream().distinct().filter(ledgerId -> {
                     try {
                         return (!flags.skipOpenLedgers || bookKeeper.isClosed(ledgerId))
                                 && !lum.isLedgerBeingReplicated(ledgerId);
@@ -172,6 +161,9 @@ public class CorrectEnsemblePlacementCommand extends
         final NavigableSet<Pair<Long, Long>> failedTargets = new ConcurrentSkipListSet<>();
         final CountDownLatch latch = new CountDownLatch(targetLedgers.size());
         for (long ledgerId : targetLedgers) {
+            if (LOG.isDebugEnabled()) {
+                LOG.debug("Start relocation of the ledger {}.", ledgerId);
+            }
             if (!flags.dryRun) {
                 try {
                     lum.acquireUnderreplicatedLedger(ledgerId);
@@ -188,103 +180,18 @@ public class CorrectEnsemblePlacementCommand extends
                         LOG.warn("Failed to open ledger {}", ledgerId);
                         return;
                     }
-
-                    final LedgerMetadata ledgerMeta = lh.getLedgerMetadata();
-                    final Map<Long, Long> ledgerFragmentsRange = new HashMap<>();
-                    Long curEntryId = null;
-                    for (Map.Entry<Long, ? extends List<BookieId>> entry :
-                            ledgerMeta.getAllEnsembles().entrySet()) {
-                        if (curEntryId != null) {
-                            ledgerFragmentsRange.put(curEntryId, entry.getKey() - 1);
-                        }
-                        curEntryId = entry.getKey();
-                    }
-                    if (curEntryId != null) {
-                        ledgerFragmentsRange.put(curEntryId, lh.getLastAddConfirmed());
-                    }
-
-                    for (Map.Entry<Long, ? extends List<BookieId>> entry : ledgerMeta.getAllEnsembles().entrySet()) {
-                        if (placementPolicy.isEnsembleAdheringToPlacementPolicy(entry.getValue(),
-                                ledgerMeta.getWriteQuorumSize(), ledgerMeta.getAckQuorumSize())
-                                == EnsemblePlacementPolicy.PlacementPolicyAdherence.FAIL) {
-                            try {
-                                final List<BookieId> currentEnsemble =  entry.getValue();
-                                // Currently, don't consider quarantinedBookies
-                                final EnsemblePlacementPolicy.PlacementResult<List<BookieId>> placementResult =
-                                        placementPolicy.replaceToAdherePlacementPolicy(
-                                                ledgerMeta.getEnsembleSize(),
-                                                ledgerMeta.getWriteQuorumSize(),
-                                                ledgerMeta.getAckQuorumSize(),
-                                                Collections.emptySet(),
-                                                currentEnsemble);
-
-                                if (placementResult.isAdheringToPolicy()
-                                        == EnsemblePlacementPolicy.PlacementPolicyAdherence.FAIL) {
-                                    LOG.warn("Failed to relocate the ensemble. So, skip the operation."
-                                                    + " ledgerId: {}, fragmentIndex: {}",
-                                            ledgerId, entry.getKey());
-                                    failedTargets.add(Pair.of(ledgerId, entry.getKey()));
-                                } else {
-                                    final List<BookieId> newEnsemble = placementResult.getResult();
-                                    final Map<Integer, BookieId> replaceBookiesMap = IntStream
-                                            .range(0, ledgerMeta.getEnsembleSize()).boxed()
-                                            .filter(i -> !newEnsemble.get(i).equals(currentEnsemble.get(i)))
-                                            .collect(Collectors.toMap(Functions.identity(), newEnsemble::get));
-                                    if (replaceBookiesMap.isEmpty()) {
-                                        LOG.warn("Failed to get bookies to replace. So, skip the operation."
-                                                        + " ledgerId: {}, fragmentIndex: {}",
-                                                ledgerId, entry.getKey());
-                                        failedTargets.add(Pair.of(ledgerId, entry.getKey()));
-                                    } else if (flags.dryRun) {
-                                        LOG.info("Would replace the ensemble. ledgerId: {}, fragmentIndex: {},"
-                                                        + " currentEnsemble: {} replaceBookiesMap {}",
-                                                ledgerId, entry.getKey(),
-                                                currentEnsemble, replaceBookiesMap);
-                                    } else {
-                                        if (LOG.isDebugEnabled()) {
-                                            LOG.debug("Try to replace the ensemble. ledgerId: {}, fragmentIndex: {},"
-                                                            + " replaceBookiesMap {}",
-                                                    ledgerId, entry.getKey(), replaceBookiesMap);
-                                        }
-                                        final LedgerFragment fragment = new LedgerFragment(lh, entry.getKey(),
-                                                ledgerFragmentsRange.get(entry.getKey()), replaceBookiesMap.keySet());
-
-                                        try {
-                                            admin.replicateLedgerFragment(lh, fragment, replaceBookiesMap,
-                                                    (lId, eId) -> {
-                                                        // This consumer is already accepted before the method returns
-                                                        // void. Therefore, use failedTargets in this consumer.
-                                                        LOG.warn("Failed to read entry {}:{}", lId, eId);
-                                                        failedTargets.add(Pair.of(ledgerId, entry.getKey()));
-                                                    });
-                                            LOG.info("Operation finished in the ensemble. ledgerId: {},"
-                                                            + " fragmentIndex: {}, replaceBookiesMap {}",
-                                                    ledgerId, entry.getKey(), replaceBookiesMap);
-                                        } catch (BKException | InterruptedException e) {
-                                            LOG.warn("Failed to replicate ledger fragment.", e);
-                                            failedTargets.add(Pair.of(ledgerId, entry.getKey()));
-                                        }
-                                    }
-                                }
-                            } catch (UnsupportedOperationException e) {
-                                LOG.warn("UnsupportedOperationException caught. The placement policy might not support"
-                                        + " replaceToAdherePlacementPolicy method.", e);
-                                failedTargets.add(Pair.of(ledgerId, entry.getKey()));
-                            }
-                        } else {
-                            if (LOG.isDebugEnabled()) {
-                                LOG.debug("The fragment is adhering to placement policy. So, skip the operation."
-                                        + " ledgerId: {}, fragmentIndex: {}", ledgerId, entry.getKey());
-                            }
-                        }
-                    }
+                    admin.relocateLedgerToAdherePlacementPolicy(lh, flags.dryRun)
+                            .forEach(e -> failedTargets.add(Pair.of(ledgerId, e)));
+                } catch (UnsupportedOperationException e) {
+                    LOG.warn("UnsupportedOperationException caught. The placement policy might not support"
+                            + " replaceToAdherePlacementPolicy method.", e);
                 } finally {
                     try {
                         if (!flags.dryRun) {
                             lum.releaseUnderreplicatedLedger(ledgerId);
                         }
                     } catch (ReplicationException e) {
-                        LOG.warn("Failed to release under replicated ledger {}.", ledgerId, e);
+                        LOG.error("Failed to release under replicated ledger {}.", ledgerId, e);
                     } finally {
                         ((CountDownLatch) ctx).countDown();
                     }
